requirements.txt
streamlit
requests
pillow
PyPDF2
google-generativeai 